wip: isolate function calculate_flexibility with basic tests
